requirements.txt
# ===== 共通 =====
Click==8.3.1
numpy==2.4.1
scikit-image==0.26.0
requests==2.32.5
six==1.17.0
Pillow==12.1.0
PyYAML==6.0.3
regex==2026.1.15
tqdm==4.67.1
opencv-python==4.13.0.92

# ===== TensorFlow 系（tag_maker 用）=====
tensorflow==2.20.0
tensorflow-io==0.31.0
tensorflow-io-gcs-filesystem==0.31.0
tensorboard==2.20.0
tensorboard-data-server==0.7.2
keras==3.13.1
h5py==3.15.1
protobuf==6.33.4
grpcio==1.76.0
absl-py==2.3.1
gast==0.7.0
opt-einsum==3.4.0
wrapt==2.0.1
flatbuffers==25.12.19
libclang==18.1.1
ml-dtypes==0.5.4
optree==0.18.0
namex==0.1.0
google-pasta==0.2.0
Markdown==3.10
Werkzeug==3.1.5
termcolor==3.3.0

# ===== Stable Diffusion XL（CUDA 12.8）=====
torch==2.9.1+cu128
torchvision==0.24.1+cu128
torchaudio==2.9.1+cu128

diffusers==0.36.0
transformers==4.57.6
accelerate==1.12.0
safetensors==0.7.0
huggingface-hub==0.36.0
tokenizers==0.22.2
peft==0.18.1

# ===== その他（依存解決用）=====
filelock==3.20.3
fsspec==2026.1.0
packaging==25.0
typing_extensions==4.15.0
networkx==3.6.1
sympy==1.14.0
jinja2==3.1.6
MarkupSafe==3.0.3
psutil==7.2.1
urllib3==2.6.3
charset-normalizer==3.4.4
certifi==2026.1.4
idna==3.11

# ===== 画像ダウンローダー（gallery-dl）=====
gallery-dl==1.23.3